Move MSAL code to AzureIdentity
