+++ examples/AzureSDKDemoSwift/Source/Common.swift
@@ -25,6 +25,7 @@
 // --------------------------------------------------------------------------
 
 import AzureCore
+import AzureIdentity
 import AzureStorageBlob
 import Foundation
 import MSAL
