+++ sdk/identity/AzureIdentity/README.md
@@ -4,18 +4,10 @@ TODO
 
 ## Getting started
 
-TODO
-
 ### Prerequisites
-
 TODO
 
 ### Install the library
-
-TODO
-
-#### Manually integrate the library into your project
-
 TODO
 
 ## Key concepts
@@ -49,4 +41,4 @@ This project has adopted the [Microsoft Open Source Code of Conduct](https://ope
 For more information see the [Code of Conduct FAQ](https://opensource.microsoft.com/codeofconduct/faq/) or contact
 [opencode@microsoft.com](mailto:opencode@microsoft.com) with any additional questions or comments.
 
-![Impressions](https://azure-sdk-impressions.azurewebsites.net/api/impressions/azure-sdk-for-ios%2Fsdk%2Fidentity%2FAzureIdentity%2FREADME.png)
+![Impressions](https://azure-sdk-impressions.azurewebsites.net/api/impressions/azure-sdk-for-ios%2Fsdk%2Fcore%2FAzureCore%2FREADME.png)
